+++ NEWS.md
@@ -1,5 +1,8 @@
 # usethis (development version)
 
+* `use_pkgdown()` and `use_logo()` now nudge towards building favicons for your website if you
+  use pkgdown (@olivroy, #1995).
+
 * The `ui_*()` functions have been marked as
   [superseded](https://lifecycle.r-lib.org/articles/stages.html#superseded).
   External users of these functions are encouraged to use the